+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - PR #829: C++: enable nvtx ranges
 - PR #837: treelite for decision forest exchange format
 - PR #870: make_blobs python function
+- PR #???: import of treelite models to FIL
 
 ## Improvements
 - PR #822: build: build.sh update to club all make targets together